Finished lesson no. 8

+++ LessonEight/src/Smartphone.java
@@ -1,8 +1,8 @@
 public class Smartphone extends Computer
 {
-    float screenDiagonal;
+    double screenDiagonal;
 
-    public Smartphone(String brandOfCPU, int ddrOfRAM, boolean hasElectricity, float screenDiagonal) {
+    public Smartphone(String brandOfCPU, int ddrOfRAM, boolean hasElectricity, double screenDiagonal) {
         super(brandOfCPU, ddrOfRAM, hasElectricity);
         this.screenDiagonal = screenDiagonal;
     }
